refactored execute
