+++ pyclesperanto_prototype/_tier0/_opencl_backend.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from ._device import get_device
 from ._pycl import OCLArray, assert_supported_ndarray_type
-
+from ._opencl_execute import execute
 
 def opencl_backend():
     return OpenCLBackend()
@@ -17,8 +17,10 @@ class OpenCLBackend():
     def asarray(self, image):
         return np.asarray(image)
 
-    @classmethod
-    def empty(cls, shape, dtype=np.float32):
+    def empty(self, shape, dtype=np.float32):
         assert_supported_ndarray_type(dtype)
         queue = get_device().queue
         return OCLArray(queue, shape, dtype)
+
+    def execute(self, anchor, opencl_kernel_filename, kernel_name, global_size, parameters, constants = None):
+        return execute(anchor, opencl_kernel_filename, kernel_name, global_size, parameters, constants)
